backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
coloredlogs==15.0.1
fastapi==0.115.8
flatbuffers==25.2.10
h11==0.14.0
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.2
onnxruntime==1.20.1
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.3
typing_extensions==4.12.2
uvicorn==0.34.0
